Add user update functionality and tests

+++ src/models/mongo/user-mongo-store.ts
@@ -1,18 +1,20 @@
 import { Types } from "mongoose";
-import { UserStore, User, NewUser } from "../store-types.js";
+import { UserStore, User, NewUser, Name, Email, Password } from "../store-types.js";
 import { UserMongoose } from "./user.js";
 
 function convertLeanUserToUser(user: Record<string, any>) {
-  user._id = String(user._id)
-  delete user.__v
-  return user as User
+  user._id = String(user._id);
+  delete user.__v;
+  return user as User;
 }
 
 export const userMongoStore: UserStore = {
   async getAllUsers(): Promise<User[]> {
     const leanUsers = await UserMongoose.find().lean();
-    const users: User[] = []
-    leanUsers.forEach(leanUser => { users.push(convertLeanUserToUser(leanUser)) });
+    const users: User[] = [];
+    leanUsers.forEach((leanUser) => {
+      users.push(convertLeanUserToUser(leanUser));
+    });
     return users;
   },
 
@@ -23,7 +25,7 @@ export const userMongoStore: UserStore = {
     }
     const leanUser = await UserMongoose.findOne({ _id: id }).lean();
     if (leanUser === null) {
-      return null
+      return null;
     }
     const user = convertLeanUserToUser(leanUser);
     return user;
@@ -40,21 +42,57 @@ export const userMongoStore: UserStore = {
   async getUserByEmail(email: string): Promise<User | null> {
     const user = await UserMongoose.findOne({ email: email }).lean();
     if (user === null) {
-      return null
+      return null;
     }
     return convertLeanUserToUser(user);
   },
 
-  async deleteUserById(id) {
-    try {
-      await UserMongoose.deleteOne({ _id: id });
-    } catch (error) {
+  async deleteUserById(id): Promise<void>  {
+    if (!Types.ObjectId.isValid(id)) {
       console.error(`Bad ID: "${id}"`);
+      return;
     }
+    await UserMongoose.deleteOne({ _id: id });
   },
 
-  async deleteAll() {
+  async deleteAll(): Promise<void> {
     await UserMongoose.deleteMany({});
+  },
+
+  async updateName(id: string, name: Name): Promise<User | null> {
+    if (!Types.ObjectId.isValid(id)) {
+      console.error(`Bad ID: "${id}"`);
+      return null;
+    }
+    const user = UserMongoose.findByIdAndUpdate(id, name, { new: true, lean: true });
+    if (user === null) {
+      return null;
+    }
+    return convertLeanUserToUser(user);
+  },
+
+  async updateEmail(id: string, email: Email): Promise<User | null> {
+    if (!Types.ObjectId.isValid(id)) {
+      console.error(`Bad ID: "${id}"`);
+      return null;
+    }
+    const user = UserMongoose.findByIdAndUpdate(id, { email: email }, { new: true, lean: true });
+    if (user === null) {
+      return null;
+    }
+    return convertLeanUserToUser(user);
+  },
+
+  async updatePassword(id: string, password: Password): Promise<User | null> {
+    if (!Types.ObjectId.isValid(id)) {
+      console.error(`Bad ID: "${id}"`);
+      return null;
+    }
+    const user = UserMongoose.findByIdAndUpdate(id, { password: password }, { new: true, lean: true });
+    if (user === null) {
+      return null;
+    }
+    return convertLeanUserToUser(user);
   },
 
   async addScope(id: string, scope: string): Promise<void> {
@@ -62,6 +100,6 @@ export const userMongoStore: UserStore = {
       console.error(`Bad ID: "${id}"`);
       return;
     }
-    await UserMongoose.findByIdAndUpdate(id, { $addToSet: { scope: scope } })
-  }
+    await UserMongoose.findByIdAndUpdate(id, { $addToSet: { scope: scope } });
+  },
 };
